requirements.txt
Pillow==3.4.1
Gooey==0.9.2.3